+++ GANDLF/schedulers/wrap_torch.py
@@ -49,9 +49,9 @@ def base_triangle(parameters):
     This function parses the parameters from the config file and returns the appropriate object
     """
 
-    # pick defaults
-    parameters["scheduler"]["min_lr"] = parameters["scheduler"].get("min_lr", 10**-3)
-    parameters["scheduler"]["max_lr"] = parameters["scheduler"].get("max_lr", 1)
+    # # pick defaults
+    # parameters["scheduler"]["min_lr"] = parameters["scheduler"].get("min_lr", 10**-3)
+    # parameters["scheduler"]["max_lr"] = parameters["scheduler"].get("max_lr", 1)
 
     clr = cyclical_lr(
         parameters["scheduler"]["step_size"],
